requirement.txt
imageio==2.31.0
lazy_loader==0.2
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pygame==2.4.0
PyWavelets==1.4.1
scikit-image==0.21.0
scipy==1.10.1
tifffile==2023.4.12
